+++ pyramid_caching/cache.py
@@ -2,6 +2,7 @@ from collections import namedtuple
 import hashlib
 import logging
 
+from pyramid.location import lineage
 from zope.interface import implementer, classImplements
 
 from pyramid_caching.events import ViewCacheHit, ViewCacheMiss
@@ -76,11 +77,13 @@ class Manager(object):
 
         if cache_content is not None:
             result = self.serializer.loads(cache_content)
+            log.debug('Cache HIT on %s', key)
             return CacheResult.hit(key, result)
         else:
             result = get_result()
             data = self.serializer.dumps(result)
             self.cache_client.add(str(key), data)
+            log.debug('Cache MISS on %s', key)
             return CacheResult.miss(key, result)
 
 
@@ -168,8 +171,8 @@ class cache_factory(object):
     ::
        @view_config(
            decorator=cache_factory(
-               predicates=[
-                   QueryStringPredicate(['name']),
+               varies_on=[
+                   QueryStringKeyModifier(['name']),
                    ],
                depends_on=[
                    RouteDependency(User, {'user_id': 'id'}),
@@ -181,22 +184,22 @@ class cache_factory(object):
 
     """
 
-    def __init__(self, predicates=None, depends_on=None):
-        self.predicates = predicates
+    def __init__(self, varies_on=None, depends_on=None):
+        self.varies_on = varies_on
         self.depends_on = depends_on
 
     def __call__(self, view):
         return ViewCacheDecorator(view,
-                                  predicates=self.predicates,
+                                  varies_on=self.varies_on,
                                   depends_on=self.depends_on,
                                   )
 
 
 class ViewCacheDecorator(object):
 
-    def __init__(self, view, predicates=None, depends_on=None):
+    def __init__(self, view, varies_on=None, depends_on=None):
         self.view = view
-        self.predicates = predicates or []
+        self.varies_on = varies_on or []
         self.depends_on = depends_on or []
 
     def __call__(self, context, request):
@@ -208,15 +211,17 @@ class ViewCacheDecorator(object):
             response.headers['X-View-Cache'] = 'DISABLED'
             return response
 
-        if not request.registry.settings['caching.enabled']:
+        if not request.registry.settings.get('caching.enabled', False):
             return nocache_result()
 
         cache_manager = request.cache_manager
 
         prefixes = [self.view.__module__, self.view.__name__]
-        prefixes.extend(self.get_predicates(request))
+        if context is not None and context.__name__ is not None:
+            prefixes.extend([context.__module__, context.__name__])
+        prefixes.extend(self.get_modifiers(request))
 
-        dependencies = self.get_dependencies(request)
+        dependencies = self.get_dependencies(context, request)
 
         try:
             result = cache_manager.get_or_cache(get_result,
@@ -236,27 +241,59 @@ class ViewCacheDecorator(object):
         response.headers['ETag'] = result.key_hash()
         return response
 
-    def get_predicates(self, request):
-        return [pred(request) for pred in self.predicates]
+    def get_modifiers(self, request):
+        return [pred(request) for pred in self.varies_on]
 
-    def get_dependencies(self, request):
-        return [dep(request) for dep in self.depends_on]
+    def get_dependencies(self, context, request):
+        return [dep(context, request) for dep in self.depends_on]
 
 
-class CollectionDependency(object):
+class ModelDependency(object):
 
-    """Model dependency on the collection of all items."""
+    """Simple dependency on a model, such as a collection."""
 
     def __init__(self, model_class):
         self.model_class = model_class
 
-    def __call__(self, request):
+    def __call__(self, context, request):
         return self.model_class
+
+
+class ResourceDependency(object):
+
+    """Dependency on a location-aware resource."""
+
+    def _is_root(self, context):
+        return context.__parent__ is None
+
+    def _is_composite(self, resource):
+        if hasattr(resource, '__composite__'):
+            return resource.__composite__
+        else:
+            return False
+
+    def _find_composite(self, resource):
+        """Walk the resource lineage to find composite resources."""
+        for location in lineage(resource):
+            if self._is_composite(location.__parent__):
+                yield location.__parent__.__name__, location.__name__
+
+    def __call__(self, context, request):
+        deps = []
+        if self._is_root(context):
+            deps.append('__root__')
+        else:
+            if self._is_composite(context):
+                deps.append(context.__name__)
+            parent_deps = dict(self._find_composite(context))
+            if parent_deps:
+                deps.append(parent_deps)
+        return tuple(deps)
 
 
 class RouteDependency(object):
 
-    """Model dependency on route segment.
+    """Model dependency based on matched route segments.
 
     Use to specify the relationships between the matched route segment elements
     and the column names of the model primary keys.
@@ -278,28 +315,31 @@ class RouteDependency(object):
     `/users/sue`, it would generate:
 
     ::
-       identity_from_request(request) --> {'id': 'sue'}
+       dependency(request) --> {'id': 'sue'}
     """
 
     def __init__(self, model_class, primary_key_elements):
         self.model_class = model_class
         self.primary_key_elements = primary_key_elements
 
-    def __call__(self, request):
+    def __call__(self, context, request):
         return (self.model_class,
                 dict((pk, request.matchdict[element])
                      for element, pk in self.primary_key_elements.iteritems())
                 )
 
 
-class QueryStringPredicate(object):
+class QueryStringKeyModifier(object):
 
-    """Cached resource predicate based on query string parameters."""
+    """Add elements from the request query string to the cache key prefix."""
 
-    def __init__(self, params):
+    def __init__(self, params=None):
         self.params = params
 
     def __call__(self, request):
-        return dict((p, request.params[p])
-                    for p in self.params
-                    if p in request.params)
+        if self.params is None:
+            return request.params.dict_of_lists()
+        else:
+            return dict((p, request.params.getall(p))
+                        for p in self.params
+                        if p in request.params)